Tests call insane main function + organization
The regression tests call insane.cli.main function rather than the
insane program. This makes tests faster, and it makes reporting test
coverage possible.

The old way is still there, but is not discoverable by nose. I need to
find a way to make it possible to toggle between running the insane
program or the insane's main function. I may also just strip the old way
completely. The new method offer more flexibility, but does not test the
actual entry point created by pip.

Accessory functions are moved in a separate module.
